added apply validation to filter step and folder restructure
