+++ ui/packages/ui/src/app/components/index.ts
@@ -4,3 +4,4 @@ export * from './FilterTreeComponent';
 export * from './KafkaConnectCluster';
 export * from './PageLoader';
 export * from './PageNotFound';
+export * from './formHelpers';
