+++ ui/packages/ui/src/app/pages/createConnector/connectorSteps/DataOptionsComponent.css
