+++ ui/packages/ui/src/app/pages/createConnector/connectorSteps/DataOptionsComponent.tsx
@@ -12,7 +12,7 @@ import { Form, Formik, useFormikContext } from "formik";
 import _ from "lodash";
 import * as React from "react";
 import { formatPropertyDefinitions, PropertyCategory } from "src/app/shared";
-import { FormComponent } from "../shared";
+import { FormComponent } from "../../../components/formHelpers";
 import "./DataOptionsComponent.css";
 
 export interface IDataOptionsComponentProps {
