+++ ui/packages/ui/src/app/pages/createConnector/connectorSteps/RuntimeOptionsComponent.tsx
@@ -5,7 +5,7 @@ import _ from "lodash";
 import * as React from "react";
 import { formatPropertyDefinitions, PropertyCategory } from "src/app/shared";
 import * as Yup from "yup";
-import { FormComponent } from "../shared";
+import { FormComponent } from "../../../components/formHelpers";
 
 export interface IRuntimeOptionsComponentProps {
   propertyDefinitions: ConnectorProperty[];
