--- ui/packages/ui/src/app/pages/createConnector/connectorSteps/dataOptions/index.tsx
@@ -1,1 +1,0 @@
-export * from './DataOptionsComponent'
