+++ ui/packages/ui/src/app/pages/createConnector/connectorSteps/index.ts
@@ -1,5 +1,5 @@
 export * from './ConnectorTypeStepComponent';
 export * from './FiltersStepComponent';
 export * from './ConfigureConnectorTypeComponent'
-export * from './dataOptions'
-export * from './runtimeOptions'
+export * from './DataOptionsComponent'
+export * from './RuntimeOptionsComponent'
