--- ui/packages/ui/src/app/pages/createConnector/connectorSteps/runtimeOptions/index.tsx
@@ -1,1 +1,0 @@
-export * from './RuntimeOptionsComponent'
